Fix engine bugs: mark placement, macro case, bounds safety
- Engine.cpp: fix rule 3.1 end-consonant check — CHR(VSI+2) was compared
  to both KEY_C and KEY_H/KEY_N simultaneously (always false); change
  second comparand to CHR(VSI+3) so "iêch/iênh/iêng" mark placement works
- Engine.cpp: fix rule 3.2 vowel-cluster mark placement — CHR(VSI) was
  compared to both KEY_I/KEY_Y/KEY_U and KEY_A at the same index (always
  false dead code); change to CHR(VSI+1) so "ia/ya/ua" mark is placed
  correctly under modern orthography
- Engine.cpp: guard _stateIndex-- with SafeDecrementIndex to prevent
  unsigned-byte underflow (wrapping to 255) on vRestore path
- Engine.cpp: remove dead commented-out debug prints
- Macro.cpp: fix modifyCaseUnicode uppercase path — code &= CAPS_MASK
  destroyed the key code; use code |= CAPS_MASK instead
- Macro.cpp: remove double semicolon ;; on return statement
- Macro.cpp: cap macroText size field at 255 (single-byte storage max)
- SmartSwitchKey.cpp: add bounds checks before reading bundleId from
  binary data to prevent out-of-bounds read on corrupt input
- ConvertTool.cpp: add missing <algorithm> include; cast break-code
  search value to Uint8 to resolve std::find type-mismatch compile error
- test_engine.cpp: full rewrite — replace calls to non-existent
  initEngine()/getTableCode() with real API; add 17 tests covering
  Telex basic, modern orthography mark placement, VNI, word break,
  Quick Telex, delete, and spelling check (all pass)
- .gitignore: exclude Linux build artifacts (*.o, test_engine binary)

+++ Sources/XoneKey/engine/ConvertTool.cpp
@@ -7,6 +7,7 @@
 //
 #include <locale>
 #include <codecvt>
+#include <algorithm>
 #include "ConvertTool.h"
 #include "Engine.h"
 #include <iostream>
@@ -166,7 +167,7 @@ string convertUtil(const string& sourceString) {
                 shouldUpperCase = true;
         } else if (t == ' ' && convertToolToCapsEachWord) {
             shouldUpperCase = true;
-        } else if (std::find(_breakCode.begin(), _breakCode.end(), t) != _breakCode.end()) {
+        } else if (std::find(_breakCode.begin(), _breakCode.end(), (Uint8)t) != _breakCode.end()) {
             hasBreak = true;
         } else {
             shouldUpperCase = false;
